requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
discord.py==1.3.2
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
typing-extensions==3.7.4.1
websockets==6.0
yarl==1.3.0
